Compile on windows
[git-p4: depot-paths = "//depot/qt/main/": change = 29207]

+++ tests/qdoc2/resolver.cpp
@@ -41,7 +41,7 @@ QString Resolver::href( const QString& name, const QString& text ) const
 {
     int k = name.find( QChar('(') );
     if ( k == -1 || (name.length() >= 10 &&
-		     name[name.length() - 3] == QChar('r') &&
+		     name[ (int)name.length() - 3] == QChar('r') &&
 		     name.right(10) == *opParenParen) )
 	k = name.length();
